Fix some C4063 warnings

+++ DirectXTex/DirectXTexD3D11.cpp
@@ -222,7 +222,7 @@ bool DirectX::IsSupportedTexture(
     const size_t iWidth = metadata.width;
     const size_t iHeight = metadata.height;
 
-    switch (fmt)
+    switch (static_cast<int>(fmt))
     {
     case DXGI_FORMAT_BC4_TYPELESS:
     case DXGI_FORMAT_BC4_UNORM:
